# Conflicts:
#	.idea/sonarlint/issuestore/1/f/1ffe51e381a6507ee647f29f84b178f41cde94f1
#	.idea/sonarlint/issuestore/5/0/509372fe1e8d433530344544d2e0334f975b6718
#	.idea/sonarlint/issuestore/7/3/7373af2083b0fe381ac5930e89e6a3438dc27bf2
#	.idea/sonarlint/issuestore/7/7/776d13b7ca88e06c982da234fb74c160d5ce7cd1
#	.idea/sonarlint/issuestore/a/7/a7d8c5ff702866ba004655188e46b0c8d28582a0
#	.idea/sonarlint/issuestore/b/a/bab5b776ce99075c2503c85880fc1b1e52be9915
#	.idea/sonarlint/issuestore/d/4/d412ae4c9ad5255a274c04060709a9a7dd03f7f7
#	.idea/sonarlint/issuestore/index.pb
#	.idea/sonarlint/securityhotspotstore/index.pb

+++ app/src/main/java/app/vibecast/presentation/weather/WeatherModel.kt
@@ -32,7 +32,7 @@ data class WeatherModel (
         val cloudCover: Int,
         val windSpeed: Double,
         val weatherConditions: List<WeatherCondition>,
-        val chanceOfRain: Double
+        val chanceOfRain: Int
     )
 
     data class WeatherCondition(
